# docs/swagger.yaml
definitions:
...
  model.Duration:
    enum:
    - -9223372036854775808
    - 9223372036854775807
    - 1
    - 1000
    - 1000000
    - 1000000000
    - 60000000000
    - 3600000000000
    - -9223372036854775808
    - 9223372036854775807
    - 1
    - 1000
    - 1000000
    - 1000000000
    - 60000000000
    - 3600000000000
    format: int64
    type: integer
    x-enum-varnames:
    - minDuration
    - maxDuration
    - Nanosecond
    - Microsecond
    - Millisecond
    - Second
    - Minute
    - Hour
    - minDuration
    - maxDuration
    - Nanosecond
    - Microsecond
    - Millisecond
    - Second
    - Minute
    - Hour
  model.HistoricalStates:
    properties:
      next_state:
        allOf:
        - $ref: '#/definitions/model.State'
        description: First state succeeding the selected time frame.
      prev_state:
        allOf:
        - $ref: '#/definitions/model.State'
        description: Last state preceding the selected time frame.
      states:
        description: All states within the selected time frame.
        items:
          $ref: '#/definitions/model.State'
        type: array
    type: object
  model.QueryBase:
    properties:
      ids:
        description: IDs for witch states are to be retrieved.
        items:
          type: string
        type: array
    type: object
  model.QueryHistorical:
    properties:
      ids:
        description: IDs for witch states are to be retrieved.
        items:
          type: string
        type: array
      range:
        allOf:
        - $ref: '#/definitions/model.Duration'
        description: Time range e.g. 24h, valid units are "ns", "us" (or "µs"), "ms",
          "s", "m", "h".
      since:
        description: Timestamp in RFC 3339 format, can be combined with 'range' or
          'until'.
        type: string
      until:
        description: Timestamp in RFC 3339 format, can be combined with 'range' or
          'since'.
        type: string
    type: object
  model.ResourceCurrentState:
    properties:
      connected:
        type: boolean
      id:
        type: string
    type: object
  model.ResourceHistoricalStates:
    properties:
      id:
        type: string
      next_state:
        allOf:
        - $ref: '#/definitions/model.State'
        description: First state succeeding the selected time frame.
      prev_state:
        allOf:
        - $ref: '#/definitions/model.State'
        description: Last state preceding the selected time frame.
      states:
        description: All states within the selected time frame.
        items:
          $ref: '#/definitions/model.State'
        type: array
    type: object
  model.State:
    properties:
      connected:
        type: boolean
      time:
        description: Timestamp in RFC 3339 format.
        type: string
    type: object
...
info:
  contact: {}
  description: Provides HTTP-API to request current and historical connection log.
  license:
    name: Apache-2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Connection Log API
  version: '{version}'
paths:
  /current/devices/{id}:
    get:
      description: Get the current state of a device.
      parameters:
      - description: device id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: device state
          schema:
            $ref: '#/definitions/model.ResourceCurrentState'
        "400":
          description: error message
          schema:
            type: string
        "401":
          description: error message
          schema:
            type: string
        "500":
          description: error message
          schema:
            type: string
      security:
      - Bearer: []
      summary: Get current device state
      tags:
      - Current states
  /current/gateways/{id}:
    get:
      description: Get the current state of a gateway.
      parameters:
      - description: gateway id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: gateway state
          schema:
            $ref: '#/definitions/model.ResourceCurrentState'
        "400":
          description: error message
          schema:
            type: string
        "401":
          description: error message
          schema:
            type: string
        "500":
          description: error message
          schema:
            type: string
      security:
      - Bearer: []
      summary: Get current gateway state
      tags:
      - Current states
  /current/query/list:
    post:
      consumes:
      - application/json
      description: 'Query current states for multiple IDs (supported: devices, gateways/hubs).'
      parameters:
      - description: query object
        in: body
        name: query
        required: true
        schema:
          $ref: '#/definitions/model.QueryBase'
      produces:
      - application/json
      responses:
        "200":
          description: current states
          schema:
            items:
              $ref: '#/definitions/model.ResourceCurrentState'
            type: array
        "400":
          description: error message
          schema:
            type: string
        "500":
          description: error message
          schema:
            type: string
      security:
      - Bearer: []
      summary: Query current states
      tags:
      - Current states
  /current/query/map:
    post:
      consumes:
      - application/json
      description: 'Query current states for multiple IDs (supported: devices, gateways/hubs).'
      parameters:
      - description: query object
        in: body
        name: query
        required: true
        schema:
          $ref: '#/definitions/model.QueryBase'
      produces:
      - application/json
      responses:
        "200":
          description: current states mapped to IDs
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: error message
          schema:
            type: string
        "500":
          description: error message
          schema:
            type: string
      security:
      - Bearer: []
      summary: Query current states
      tags:
      - Current states
  /historical/devices/{id}:
    get:
      description: Get the historical states of a device.
      parameters:
      - description: device id
        in: path
        name: id
        required: true
        type: string
      - description: time range e.g. 24h, valid units are 'ns', 'us' (or 'µs'), 'ms',
          's', 'm', 'h'
        in: query
        name: range
        type: string
      - description: timestamp in RFC 3339 format, can be combined with 'range' or
          'until'
        in: query
        name: since
        type: string
      - description: timestamp in RFC 3339 format, can be combined with 'range' or
          'since'
        in: query
        name: until
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: device state
          schema:
            $ref: '#/definitions/model.ResourceHistoricalStates'
        "400":
          description: error message
          schema:
            type: string
        "401":
          description: error message
          schema:
            type: string
        "500":
          description: error message
          schema:
            type: string
      security:
      - Bearer: []
      summary: Get historical device states
      tags:
      - Historical states
  /historical/gateways/{id}:
    get:
      description: Get the historical states of a gateway.
      parameters:
      - description: gateway id
        in: path
        name: id
        required: true
        type: string
      - description: time range e.g. 24h, valid units are 'ns', 'us' (or 'µs'), 'ms',
          's', 'm', 'h'
        in: query
        name: range
        type: string
      - description: timestamp in RFC 3339 format, can be combined with 'range' or
          'until'
        in: query
        name: since
        type: string
      - description: timestamp in RFC 3339 format, can be combined with 'range' or
          'since'
        in: query
        name: until
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: gateway states
          schema:
            $ref: '#/definitions/model.ResourceHistoricalStates'
        "400":
          description: error message
          schema:
            type: string
        "401":
          description: error message
          schema:
            type: string
        "500":
          description: error message
          schema:
            type: string
      security:
      - Bearer: []
      summary: Get historical gateway states
      tags:
      - Historical states
  /historical/query/list:
    post:
      consumes:
      - application/json
      description: 'Query current historical states for multiple IDs (supported: devices,
        gateways/hubs).'
      parameters:
      - description: query object
        in: body
        name: query
        required: true
        schema:
          $ref: '#/definitions/model.QueryHistorical'
      produces:
      - application/json
      responses:
        "200":
          description: historical states
          schema:
            items:
              $ref: '#/definitions/model.ResourceHistoricalStates'
            type: array
        "400":
          description: error message
          schema:
            type: string
        "500":
          description: error message
          schema:
            type: string
      security:
      - Bearer: []
      summary: Query historical states
      tags:
      - Historical states
  /historical/query/map:
    post:
      consumes:
      - application/json
      description: 'Query current historical states for multiple IDs (supported: devices,
        gateways/hubs).'
      parameters:
      - description: query object
        in: body
        name: query
        required: true
        schema:
          $ref: '#/definitions/model.QueryHistorical'
      produces:
      - application/json
      responses:
        "200":
          description: historical states mapped to IDs
          schema:
            additionalProperties:
              $ref: '#/definitions/model.HistoricalStates'
            type: object
        "400":
          description: error message
          schema:
            type: string
        "500":
          description: error message
          schema:
            type: string
      security:
      - Bearer: []
      summary: Query historical states
      tags:
      - Historical states
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"